# Conflicts:
#	modules/decima-ext-model-exporter/src/main/java/com/shade/decima/ui/data/viewer/model/dmf/DMFExporter.java

+++ modules/decima-model/src/main/java/com/shade/decima/model/packfile/resource/BufferResource.java
@@ -17,7 +17,7 @@ public class BufferResource implements Resource {
     }
 
     @Override
-    public long read(@NotNull ByteBuffer buffer) throws IOException {
+    public int read(@NotNull ByteBuffer buffer) throws IOException {
         if (position < 0) {
             throw new IOException("Resource is closed");
         }
